+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master
 
+* Add Android Kotlin templates.  Use `-t android-kotlin` to use it.  
+  [Xavier F. Gouchet](https://github.com/xgouchet)
+  [#42](https://github.com/NijiDigital/gyro/pull/42)
+  
+## 1.2.0
+
 * Now the swift3 template is compatible swift4 too, so you can use -t swift3 even for Swift 4.  
   [Jérémie GOAS](https://github.com/jgoas)
   [#41](https://github.com/NijiDigital/gyro/pull/41)